Moved Carousel component to "shared" folder.
